Added typings and Typed tests
These tests won' compile because the typings are incomplete, but they're a godo guide to work on. As types are added, more of these should resolve.

Some tests are for typechecking or checking errors are thrown when types are incomplete--these are not necessary to be satisfied so we should remove these, and ensure our types enforce the condition that the test specifies.